refactor : repository id type 수정 - TeacherRepository : Long -> Integer - StudentRepository : Long -> Integer - 디렉토리 rename : domain -> entity
